requirements.txt
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
aiohttp>=3.9,<4.0
aiosqlite>=0.19,<1.0
aiodns>=3.0,<4.0
fuzzywuzzy>=0.18,<1.0
python-Levenshtein>=0.21,<1.0
python-dotenv>=1.0,<2.0
paramiko>=3.4,<4.0
